Se agregaron los dos modelos adicionales que se utilizaran en las tablas
