+++ .idea/misc.xml
@@ -3,7 +3,8 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-        <entry key="..\:/Users/INTEL/Documents/ProyectoFinalDispositivosMoviles/app/src/main/res/layout/activity_bottom.xml" value="0.13333333333333333" />
+        <entry key="..\:/Users/INTEL/Documents/ProyectoFinalDispositivosMoviles/app/src/main/res/drawable/ic_video_game.xml" value="0.19583333333333333" />
+        <entry key="..\:/Users/INTEL/Documents/ProyectoFinalDispositivosMoviles/app/src/main/res/layout/activity_bottom.xml" value="0.1" />
         <entry key="..\:/Users/INTEL/Documents/ProyectoFinalDispositivosMoviles/app/src/main/res/layout/activity_login.xml" value="0.3333333333333333" />
         <entry key="..\:/Users/INTEL/Documents/ProyectoFinalDispositivosMoviles/app/src/main/res/layout/activity_main.xml" value="0.3333333333333333" />
         <entry key="..\:/Users/INTEL/Documents/ProyectoFinalDispositivosMoviles/app/src/main/res/layout/activity_register.xml" value="0.22" />
@@ -11,7 +12,9 @@
         <entry key="..\:/Users/INTEL/Documents/ProyectoFinalDispositivosMoviles/app/src/main/res/layout/fragment_dashboard.xml" value="0.1" />
         <entry key="..\:/Users/INTEL/Documents/ProyectoFinalDispositivosMoviles/app/src/main/res/layout/fragment_home.xml" value="0.1796875" />
         <entry key="..\:/Users/INTEL/Documents/ProyectoFinalDispositivosMoviles/app/src/main/res/layout/fragment_profile.xml" value="0.22" />
-        <entry key="..\:/Users/INTEL/Documents/ProyectoFinalDispositivosMoviles/app/src/main/res/layout/fragment_scores.xml" value="0.22" />
+        <entry key="..\:/Users/INTEL/Documents/ProyectoFinalDispositivosMoviles/app/src/main/res/layout/fragment_scores.xml" value="0.3333333333333333" />
+        <entry key="..\:/Users/INTEL/Documents/ProyectoFinalDispositivosMoviles/app/src/main/res/layout/fragment_starting_group.xml" value="0.1" />
+        <entry key="..\:/Users/INTEL/Documents/ProyectoFinalDispositivosMoviles/app/src/main/res/layout/fragment_starting_selection.xml" value="0.1" />
         <entry key="..\:/Users/INTEL/Documents/ProyectoFinalDispositivosMoviles/app/src/main/res/menu/bottom_nav_menu.xml" value="0.13978494623655913" />
         <entry key="..\:/Users/INTEL/Documents/ProyectoFinalDispositivosMoviles/app/src/main/res/menu/menu_overflow.xml" value="0.22" />
         <entry key="..\:/Users/INTEL/Documents/ProyectoFinalDispositivosMoviles/app/src/main/res/menu/overflow_menu.xml" value="0.1796875" />
