+++ app/src/main/java/com/migue/proyectofinal/models/Game.kt
@@ -1,3 +1,13 @@
 package com.migue.proyectofinal.models
 
-data class Games()
+import java.io.Serializable
+
+data class Game(
+    val namePlayer1: String,
+    val namePlayer2: String,
+    val socrePlayer1: String,
+    val scorePlayer2: String,
+    val winner: String,
+    val typeGame: String
+) : Serializable
+
