+++ app/src/main/java/com/migue/proyectofinal/ui/main/MainActivity.kt
@@ -19,23 +19,7 @@ class MainActivity : AppCompatActivity() {
         setContentView(mainBinding.root)
     }
 
-    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
-        menuInflater.inflate(R.menu.overflow_menu, menu)
-        return true
-    }
 
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        when(item.itemId){
-            R.id.menu_sing_out -> goToLoginActivity()
-        }
-        return super.onOptionsItemSelected(item)
-    }
-
-    private fun goToLoginActivity() {
-        val intent = Intent(this, LoginActivity::class.java)
-        intent.flags = Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_NEW_TASK
-        startActivity(intent)
-    }
 
     /*private lateinit var mainBinding: ActivityMainBinding
