+++ app/src/main/java/com/migue/proyectofinal/ui/startingselection/StartingSelectionFragment.kt
@@ -44,4 +44,6 @@ class StartingSelectionFragment : Fragment() {
             }
         }
     }
+
+
 }
